[Win32] Move DPI change handling from Control to Shell
Currently, DPI change handling is implemented in Control. However,
according OS events are only created and reasonable at Shell level.
This change thus moves the according implementations to the Shell class.
It also limits the last queued DPI event storage for cancellation on
subsequent events to the Shell class.
